Migrate Mutation.updateSchemaPolicyForOrganization
